use +- keys for zoom without cmd
This enables to keep browser zoom behavior for accessilbility
