+++ apps/designer/app/designer/features/breakpoints/use-subscribe-shortcuts.ts
@@ -3,8 +3,6 @@ import { useBreakpoints } from "~/shared/nano-states";
 import { useZoom, useSelectedBreakpoint } from "../../shared/nano-states";
 import { minZoom } from "./zoom-setting";
 import { utils } from "@webstudio-is/project";
-import { shortcuts } from "~/shared/shortcuts";
-import { useHotkeys } from "react-hotkeys-hook";
 
 export const useSubscribeSelectBreakpointFromShortcut = () => {
   const [breakpoints] = useBreakpoints();
@@ -21,16 +19,6 @@ const zoomStep = 20;
 
 export const useSubscribeZoomFromShortcut = () => {
   const [zoom, setZoom] = useZoom();
-
-  // We need to prevent browser default behavior for zooming in/out
-  // Maybe we should use shift modifier and keep the default behaivor functional?
-  useHotkeys(
-    shortcuts.zoom,
-    (event) => {
-      event.preventDefault();
-    },
-    []
-  );
 
   useSubscribe("zoom", (direction) => {
     if (direction === "zoomIn") {
